requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bcrypt==3.2.0
blinker==1.4
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
colorama==0.4.4
contourpy==1.3.1
cryptography==3.4.8
cupy-cuda12x==13.2.0
cycler==0.12.1
Cython==0.29.28
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
fasteners==0.14.1
fastrlock==0.8.2
ffmpeg-python==0.2.0
flatbuffers==24.3.25
fonttools==4.55.0
future==0.18.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
httplib2==0.20.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib-metadata==4.6.4
jeepney==0.7.1
joblib==1.4.2
keras==3.6.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
namex==0.0.8
netifaces==0.11.0
numpy==2.0.2
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
pillow==11.0.0
proglog==0.1.10
ptyprocess==0.7.0
pyasn1==0.6.1
pycairo==1.20.1
Pygments==2.18.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyRFC3339==1.1
PySimpleGUI==5.0.6
python-dateutil==2.9.0.post0
pytz==2024.2
pyxdg==0.27
reportlab==3.6.8
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wadllib==1.3.6
Werkzeug==3.1.3
wrapt==1.16.0
xdg==5
zipp==1.0.0
